requirements.txt
python-dotenv
mysql-connector-python
faker
streamlit
ortools
fpdf2
